# netlink: optimize MarshalAttributes