New benchmarks:

```
✔ ~/src/github.com/mdlayher/netlink [master|✚ 1…6] 
11:05 $ cat after.bench | grep -v Unmarshal
goos: linux
goarch: amd64
pkg: github.com/mdlayher/netlink
BenchmarkMarshalAttributes/0-8          200000000                8.32 ns/op            0 B/op          0 allocs/op
BenchmarkMarshalAttributes/1-8          50000000                31.3 ns/op             8 B/op          1 allocs/op
BenchmarkMarshalAttributes/8-8          10000000               168 ns/op              96 B/op          1 allocs/op
BenchmarkMarshalAttributes/64-8          1000000              2189 ns/op            4864 B/op          1 allocs/op
BenchmarkMarshalAttributes/512-8           30000             49191 ns/op          270336 B/op          1 allocs/op
PASS
ok      github.com/mdlayher/netlink     20.476s
goos: linux
goarch: amd64
pkg: github.com/mdlayher/netlink
BenchmarkMarshalAttributes/0-8          200000000                8.53 ns/op            0 B/op          0 allocs/op
BenchmarkMarshalAttributes/1-8          50000000                30.9 ns/op             8 B/op          1 allocs/op
BenchmarkMarshalAttributes/8-8          10000000               173 ns/op              96 B/op          1 allocs/op
BenchmarkMarshalAttributes/64-8          1000000              2183 ns/op            4864 B/op          1 allocs/op
BenchmarkMarshalAttributes/512-8           30000             51241 ns/op          270336 B/op          1 allocs/op
PASS
ok      github.com/mdlayher/netlink     20.548s
goos: linux
goarch: amd64
pkg: github.com/mdlayher/netlink
BenchmarkMarshalAttributes/0-8          200000000                8.22 ns/op            0 B/op          0 allocs/op
BenchmarkMarshalAttributes/1-8          50000000                30.9 ns/op             8 B/op          1 allocs/op
BenchmarkMarshalAttributes/8-8          10000000               166 ns/op              96 B/op          1 allocs/op
BenchmarkMarshalAttributes/64-8          1000000              2171 ns/op            4864 B/op          1 allocs/op
BenchmarkMarshalAttributes/512-8           30000             49360 ns/op          270336 B/op          1 allocs/op
PASS
ok      github.com/mdlayher/netlink     20.194s
goos: linux
goarch: amd64
pkg: github.com/mdlayher/netlink
BenchmarkMarshalAttributes/0-8          200000000                8.30 ns/op            0 B/op          0 allocs/op
BenchmarkMarshalAttributes/1-8          50000000                31.1 ns/op             8 B/op          1 allocs/op
BenchmarkMarshalAttributes/8-8          10000000               173 ns/op              96 B/op          1 allocs/op
BenchmarkMarshalAttributes/64-8          1000000              2217 ns/op            4864 B/op          1 allocs/op
BenchmarkMarshalAttributes/512-8           30000             50986 ns/op          270336 B/op          1 allocs/op
PASS
ok      github.com/mdlayher/netlink     19.953s
goos: linux
goarch: amd64
pkg: github.com/mdlayher/netlink
BenchmarkMarshalAttributes/0-8          200000000                8.37 ns/op            0 B/op          0 allocs/op
BenchmarkMarshalAttributes/1-8          50000000                31.3 ns/op             8 B/op          1 allocs/op
BenchmarkMarshalAttributes/8-8          10000000               172 ns/op              96 B/op          1 allocs/op
BenchmarkMarshalAttributes/64-8          1000000              2159 ns/op            4864 B/op          1 allocs/op
BenchmarkMarshalAttributes/512-8           30000             51100 ns/op          270336 B/op          1 allocs/op
PASS
ok      github.com/mdlayher/netlink     20.451s
```

Comparison:

```
✔ ~/src/github.com/mdlayher/netlink [mdl-attr-fastmarshal|…6] 
11:07 $ benchstat before.bench after.bench | grep -v Unmarshal
name                       old time/op    new time/op    delta
MarshalAttributes/0-8        8.35ns ± 0%    8.35ns ± 2%     ~     (p=0.643 n=5+5)
MarshalAttributes/1-8        64.0ns ± 1%    31.1ns ± 1%  -51.39%  (p=0.008 n=5+5)
MarshalAttributes/8-8         411ns ± 4%     170ns ± 3%  -58.52%  (p=0.008 n=5+5)
MarshalAttributes/64-8       6.62µs ±14%    2.18µs ± 2%  -66.99%  (p=0.008 n=5+5)
MarshalAttributes/512-8       178µs ± 4%      50µs ± 2%  -71.71%  (p=0.008 n=5+5)

name                       old alloc/op   new alloc/op   delta
MarshalAttributes/0-8         0.00B          0.00B          ~     (all equal)
MarshalAttributes/1-8         24.0B ± 0%      8.0B ± 0%  -66.67%  (p=0.008 n=5+5)
MarshalAttributes/8-8          320B ± 0%       96B ± 0%  -70.00%  (p=0.008 n=5+5)
MarshalAttributes/64-8       14.6kB ± 0%     4.9kB ± 0%  -66.67%  (p=0.008 n=5+5)
MarshalAttributes/512-8       885kB ± 0%     270kB ± 0%  -69.44%  (p=0.008 n=5+5)

name                       old allocs/op  new allocs/op  delta
MarshalAttributes/0-8          0.00           0.00          ~     (all equal)
MarshalAttributes/1-8          3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.008 n=5+5)
MarshalAttributes/8-8          10.0 ± 0%       1.0 ± 0%  -90.00%  (p=0.008 n=5+5)
MarshalAttributes/64-8         66.0 ± 0%       1.0 ± 0%  -98.48%  (p=0.008 n=5+5)
MarshalAttributes/512-8         514 ± 0%         1 ± 0%  -99.81%  (p=0.008 n=5+5)
```